widget/wcolorpickeraction: Do not hand out raw pointer to color picker
